Do not set errno in __filter_syscall*.
musl's __syscall function is expected to return the (negated) errno value as
return value. Therefore setting errno and setting the return value to -1 should
not be done as part of __filter_syscall*. This is done by the syscall_ret
function instead. Calls to syscall (not __syscall) will have set the errno
value as expected.