calloc: initialize memory correctly with zero
In dc84c0a0 an optimization was introduced to avoid zeroing out pages
in some cases. However due to an update of musl __malloc0 is gone
and the check if memory was allocated via mmap was moved to calloc.
This commit removes this check to unconditionally zero out memory again.