Do not call syscall(futex, x, FUTEX_WAIT, [...]) from __wait when in scheduling context.
Under some circumstances we have to wait for a lock to be released in the
scheduling context, e.g. in the following case: lthread_free -> free -> [...]
-> lock_bin -> lock -> __wait

With this commit, we will spin indefinitely to wait for the lock to be released
as we cannot call futex from outside of an lthread. Any lock that needs to be
hold by code executed from the scheduling context should have low lock
contention and/or should only be hold for short amounts of time. Therefore we
now busy wait for the lock to be released.