Fix: Do not call syscall(futex, x, FUTEX_WAIT, [...]) from __wait when in scheduling context.
This is a revision of the following commit which didn't fully fix the problem
as syscall(futex, [...]) could still be called in __wait from the scheduling
context under certain circumstances.

commit 904d3c3acc5423013afe5430e3d8260652c30ee0
Author: Christian Priebe <cp3213@ic.ac.uk>
Date:   Sun Jul 14 14:32:09 2019 +0100

Do not call syscall(futex, x, FUTEX_WAIT, [...]) from __wait when in scheduling context.

Under some circumstances we have to wait for a lock to be released in the
scheduling context, e.g. in the following case: lthread_free -> free -> [...]
-> lock_bin -> lock -> __wait

With this commit, we will spin indefinitely to wait for the lock to be released
as we cannot call futex from outside of an lthread. Any lock that needs to be
hold by code executed from the scheduling context should have low lock
contention and/or should only be hold for short amounts of time. Therefore we
now busy wait for the lock to be released.