Validate call IDs on ecalls.
In particular, ensure that signal handling code can only be invoked after an
asynchronous exit (indicated by CSSA > 0). Otherwise it is possible to trick
SGX-LKL into restoring register values from an uninitialised SSA.

This commit adds checks to ensure that the ecall ID is valid, i.e. is one of

0: THREAD_CREATE,
1: RESUME,
2: HANDLE_SIGNAL

If CSSA is 0, the ID must be either 0 (THREAD_CREATE) or 1 (RESUME), otherwise
it must bust be either 1 (RESUME) or 2 (HANDLE_SIGNAL).

Note: This code will be replaced with the planned migration to the Open Enclave
SDK.

Thanks to Jo Van Bulck, KU Leuven and David Oswald, University of Birmingham
for reporting this issue.